requirements.txt
absl-py==0.9.0
argon2-cffi==20.1.0
astroid @ file:///tmp/build/80754af9/astroid_1592495881661/work
astunparse==1.6.3
attrs==19.3.0
backcall==0.2.0
beautifulsoup4==4.9.1
bleach==3.1.5
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cmake==3.18.0
colorama==0.4.3
contextlib2==0.6.0.post1
cryptography==3.0
cycler==0.10.0
Cython==0.29.21
cython-bbox==0.1.3
decorator==4.4.2
defusedxml==0.6.0
dlib @ file:///home/conda/feedstock_root/build_artifacts/dlib_1597241653647/work
docutils==0.16
easydict==1.9
entrypoints==0.3
future==0.18.2
gast==0.3.3
google-auth==1.20.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
graphviz==0.8.4
grpcio==1.31.0
h5py==2.10.0
idna==2.10
importlib-metadata==1.7.0
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.17.2
jeepney==0.4.3
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras==2.4.3
Keras-Preprocessing==1.1.2
keyring==21.3.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lxml==4.5.2
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.1
mccabe==0.6.1
mistune==0.8.4
mxnet==1.6.0
natsort==7.0.1
nbconvert==5.6.1
nbformat==5.0.7
netron==4.5.0
notebook==6.1.3
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1595522867663/work
oauthlib==3.1.0
onnx==1.7.0
onnx-simplifier==0.2.16
onnxruntime==1.4.0
opencv-contrib-python==3.4.4.19
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.0
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pkginfo==1.5.0.1
prometheus-client==0.8.0
prompt-toolkit==3.0.6
protobuf==3.13.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
pylint @ file:///tmp/build/80754af9/pylint_1592496044040/work
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.6
QtPy==1.9.0
readme-renderer==26.0
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.4.1
SecretStorage==3.1.2
Send2Trash==1.5.0
six==1.15.0
slim==0.1
soupsieve==2.0.1
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
tensorflow-object-detection-api==0.1.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
toml @ file:///tmp/build/80754af9/toml_1592853716807/work
torch==1.2.0
torch-vision==0.1.6.dev0
torchvision==0.4.0
tornado==6.0.4
tqdm==4.48.2
traitlets==4.3.3
twine==3.2.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
vision==1.0.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.1.0
